modularize test differential linkages
Modularizes test differential linkages to extend statistical basis of differential cell signaling tests to fisher's exact test and logistic regression. If the logistic regression method is used, a model formula can be passed in the format of a linear model to test effects of multiple variables on signaling probability and their interaction.