initial generate_expression_bootstrap
For an approach to bootstrapping data for use with dominosignal, I have decided not to have the running of create_domino and build_domino within the scope of one full bootstrapping function as the users will not be able to edit parameters of the two essential create and build functions. Thus, I have made a function for the sampling with replacement to generate a bootstrap in the form of the expression values and cluster assignments used as inputs to create_domino.